requirements.txt
Django==4.2.10
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==4.6.0
django-cors-headers==4.2.0
channels==4.0.0
daphne==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==9.5.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
